requirements.txt
strands-agents>=1.0
strands-agents-tools>=0.1
strands-agents-evals>=0.1
mlflow>=2.15
sagemaker-mlflow
boto3>=1.34
datasets>=2.19
pandas>=2.0
textstat
evaluate
nltk
rouge-score
tiktoken
transformers
deepeval
litellm
nest_asyncio
python-docx
